tools/build_templates/jiro/requirements/dev.txt
# run 'pip install -r requirements/prod.txt'
-r common.txt
ipython
ipdb
django-extensions
Werkzeug
